[bad] fractional shares
This suffers from a lack of precision. When testing against a larger
number of transactions, it eventually differs. The original test case
used shares*100, amt-per-share/100 to achieve an integer # shares, and
the comparison used shares and amt-per-share in their original scale.

Do not use this.

+++ portfolio/summary.go
@@ -260,7 +260,7 @@ func makeAnnualGainsSummaryTxs(
 
 	// Add length of yearsWithGains to the share balance, as we'll sell one share per year
 	// This will generally always be non-zero for non-registered affiliates
-	nBaseShares := sumPostStatus.ShareBalance + uint32(len(yearsWithGains))
+	nBaseShares := sumPostStatus.ShareBalance + float64(len(yearsWithGains))
 	if nBaseShares > 0 {
 		tx := deltas[latestSummarizableDeltaIdx].Tx
 		// Get the earliest year, and use Jan 1 of the previous year for the buy.
